Update to XCTest, recommended project settings, and fix warnings in Xcode 6.

+++ external/universalchardet/src/base/nsCharSetProber.cpp
@@ -41,7 +41,7 @@ bool nsCharSetProber::FilterWithoutEnglishLetters(const char* aBuf, uint32_t aLe
   if (meetMSB && curPtr > prevPtr) 
     while (prevPtr < curPtr) *newptr++ = *prevPtr++;  
 
-  newLen = newptr - *newBuf;
+  newLen = (uint32_t)(newptr - *newBuf);
 
   return true;
 }
@@ -84,9 +84,9 @@ bool nsCharSetProber::FilterWithEnglishLetters(const char* aBuf, uint32_t aLen, 
   // and it is not inside a tag then keep it.
   if (!isInTag)
     while (prevPtr < curPtr)
-      *newptr++ = *prevPtr++;  
+      *newptr++ = *prevPtr++;
 
-  newLen = newptr - *newBuf;
+  newLen = (uint32_t)(newptr - *newBuf);
 
   return true;
 }
